ARROW-16974: [GLib] Make C99 compatible (#13512)
We can't use multiple "typedef struct _XXX XXX" in C99.

Authored-by: Sutou Kouhei <kou@clear-code.com>
