+++ c_glib/arrow-glib/scalar.cpp
@@ -19,6 +19,7 @@
 
 #include <arrow-glib/basic-array.hpp>
 #include <arrow-glib/buffer.hpp>
+#include <arrow-glib/compute.h>
 #include <arrow-glib/data-type.hpp>
 #include <arrow-glib/decimal.hpp>
 #include <arrow-glib/error.hpp>
