+++ c_glib/meson.build
@@ -20,6 +20,7 @@
 project('arrow-glib', 'c', 'cpp',
         license: 'Apache-2.0',
         default_options: [
+          'c_std=c99',
           'cpp_std=c++11',
         ])
